Make foreign key colume name back to `edgeName_id' style(reverse #286).
To achive this goal this commit does following things:

- Introduces edge.ToFrom() to define bi-directional edges(o2obidi, m2mbidi), this kind of edges set the "Bidi: true" field early to indicate the bi-directional relationship.
- Make all other kind of relationships must have To() edge and From() edge both defined and paired.
- With the two constraints introduced above, an inverse edge isn't optional anymore and can not be dropped. So the `edgeName_id' kind Foreign key name won't change.

The purpose of disallowing From() edges to be dropped is to keep FK names unchanged.

The purpose of introducing ToFrom() edge is that before this commit, a bi-directional edge looks just like an To() edge only(although to the same type as the edge owner) and has no inverse edge defined, this violates the principle described above that a To() edge must have an inverse From() edge paired.

This commit fixes #548.

I also believe that with these changes, it will be easier to add better offline/edge migration support in the future.