Add FieldDefault method to Mutation interface and entc template.
The Save/SaveX methods on the generated <T>Create and <T>Update mutation
wrappers set field defaults before hooks are run so code has no way to
tell which fields were explicitly set and which are using default
values.

This change adds a FieldDefault method to Mutation that is similar to
the existing Field method, only it returns a fields default value. This
allows code to compare the current and default value and infer if the
field was explicitly set to a non-default value.