Add `BuildFlags` to schema loader configuration
This allows passing arbitrary build flags to the schema package loader,
which in turn can be used to pass build constraints affecting what
schema types and/or methods are visible.

As a particular use case, this should make it easier to implement
two-pass code generation to deal with circular imports in hooks (#892).